fixing file change triggers to listen for creationtime since visual studio doesn't edit file..
todo: fix finding project from filepath in OnProjectItemChanged(
